ExecutorService extends Synchronization::Object
