+++ lib/concurrent/executor/java_thread_pool_executor.rb
@@ -1,12 +1,11 @@
 if Concurrent.on_jruby?
-  require_relative 'executor'
+  require 'concurrent/executor/executor_service'
 
   module Concurrent
 
     # @!macro thread_pool_executor
     # @!macro thread_pool_options
-    class JavaThreadPoolExecutor
-      include JavaExecutor
+    class JavaThreadPoolExecutor < JavaExecutorService
 
       # Default maximum number of threads that will be created in the pool.
       DEFAULT_MAX_POOL_SIZE = java.lang.Integer::MAX_VALUE # 2147483647
@@ -53,6 +52,8 @@ if Concurrent.on_jruby?
       #
       # @see http://docs.oracle.com/javase/7/docs/api/java/util/concurrent/ThreadPoolExecutor.html
       def initialize(opts = {})
+        super()
+
         min_length = opts.fetch(:min_threads, DEFAULT_MIN_POOL_SIZE).to_i
         max_length = opts.fetch(:max_threads, DEFAULT_MAX_POOL_SIZE).to_i
         idletime = opts.fetch(:idletime, DEFAULT_THREAD_IDLETIMEOUT).to_i
@@ -63,7 +64,7 @@ if Concurrent.on_jruby?
         raise ArgumentError.new('max_threads must be greater than zero') if max_length <= 0
         raise ArgumentError.new('min_threads cannot be less than zero') if min_length < 0
         raise ArgumentError.new('min_threads cannot be more than max_threads') if min_length > max_length
-        raise ArgumentError.new("#{fallback_policy} is not a valid fallback policy") unless FALLBACK_POLICIES.include?(@fallback_policy)
+        raise ArgumentError.new("#{fallback_policy} is not a valid fallback policy") unless FALLBACK_POLICY_CLASSES.include?(@fallback_policy)
 
         if @max_queue == 0
           queue = java.util.concurrent.LinkedBlockingQueue.new
@@ -74,7 +75,7 @@ if Concurrent.on_jruby?
         @executor = java.util.concurrent.ThreadPoolExecutor.new(
           min_length, max_length,
           idletime, java.util.concurrent.TimeUnit::SECONDS,
-          queue, FALLBACK_POLICIES[@fallback_policy].new)
+          queue, FALLBACK_POLICY_CLASSES[@fallback_policy].new)
 
         self.auto_terminate = opts.fetch(:auto_terminate, true)
       end
