+++ lib/concurrent/executor/ruby_single_thread_executor.rb
@@ -1,11 +1,11 @@
-require_relative 'executor'
+require 'concurrent/executor/executor'
+require 'concurrent/executor/executor_service'
 
 module Concurrent
 
   # @!macro single_thread_executor
   # @!macro thread_pool_options
-  class RubySingleThreadExecutor
-    include RubyExecutor
+  class RubySingleThreadExecutor < RubyExecutorService
     include SerialExecutor
 
     # Create a new thread pool.
@@ -18,11 +18,11 @@ module Concurrent
     # @see http://docs.oracle.com/javase/7/docs/api/java/util/concurrent/Executors.html
     # @see http://docs.oracle.com/javase/8/docs/api/java/util/concurrent/ExecutorService.html
     def initialize(opts = {})
+      super()
       @queue = Queue.new
       @thread = nil
       @fallback_policy = opts.fetch(:fallback_policy, :discard)
       raise ArgumentError.new("#{@fallback_policy} is not a valid fallback policy") unless FALLBACK_POLICIES.include?(@fallback_policy)
-      init_executor
       self.auto_terminate = opts.fetch(:auto_terminate, true)
     end
